fix empty tags
