+++ teradata/metadata.csv
@@ -1,17 +1,23 @@
 metric_name,metric_type,interval,unit_name,per_unit_name,description,orientation,integration,short_name,curated_metric
-teradata.disk_space.alloc_perm.total
-teradata.disk_space.alloc_spool.total
-teradata.disk_space.alloc_temp.total
-teradata.disk_space.curr_persist_spool.total
-teradata.disk_space.curr_temp.total
-teradata.disk_space.max_temp.total
-teradata.disk_space.peak_perm.total
-teradata.disk_space.peak_persist_spool.total
-teradata.disk_space.peak_spool.total
-teradata.disk_space.peak_temp.total
-teradata.disk_space.perm_skew.total
-teradata.disk_space.spool_skew.total
-teradata.disk_space.temp_skew.total
+teradata.disk_space.max_perm.total,gauge,,byte,,"Maximum permanent space, in bytes, that is allocated to the database on a specified AMP",0,teradata,maxperm,
+teradata.disk_space.max_spool.total,gauge,,byte,,"Maximum spool space, in bytes, that is allocated to the database on a specified AMP",0,teradata,maxspool,
+teradata.disk_space.max_temp.total,gauge,,byte,,"Maximum temporary space, in bytes, that is allocated to the database on a specified AMP",0,teradata,maxtemp,
+teradata.disk_space.curr_perm.total,gauge,,byte,,"Permanent space per AMP currently being used",0,teradata,currperm,
+teradata.disk_space.curr_spool.total,gauge,,byte,,"Spool space per AMP currently being used by the user",0,teradata,currspool,
+teradata.disk_space.curr_persist_spool.total,gauge,,byte,,"Persistent spool space per AMP currently being used by the user",0,teradata,curr_pers_spool,
+teradata.disk_space.curr_temp.total,gauge,,byte,,"The number of bytes currently used by a temporary table per virtual processor (vproc)",0,teradata,curtemp,
+teradata.disk_space.peak_perm.total,gauge,,byte,,"Maximum amount of permanent space per AMP that has been used by the database since the last time the DBC.ClearPeakDisk macro was run",0,teradata,peakperm,
+teradata.disk_space.peak_spool.total,gauge,,byte,,"Maximum spool space, in bytes, that was used at any one time by the database on a specified AMP",0,teradata,peakspool,
+teradata.disk_space.peak_persist_spool.total,gauge,,byte,,"Maximum persistent spool space, in bytes, that was used at any one time by the database on a specified AMP",0,teradata,peak_pers_spool,
+teradata.disk_space.peak_temp.total,gauge,,byte,,"Peak number of bytes used at one time by a temporary table per vproc",0,teradata,peaktemp,
+teradata.disk_space.max_prof_spool.total,gauge,,byte,,"the profile SPOOL space limit per AMP for the user if the user is assigned a profile which has a SPOOL space setting",0,teradata,max_prof_spool,
+teradata.disk_space.max_prof_temp.total,gauge,,byte,,"the profile TEMPORARY space limit per AMP for the user if the user is assigned a profile which has a TEMPORARY space setting",0,teradata,max_prof_temp,
+teradata.disk_space.alloc_perm.total,gauge,,byte,,"current total allocation in bytes for permanent and user defined temporary table space",0,teradata,allocperm,
+teradata.disk_space.alloc_spool.total,gauge,,byte,,"current total allocation in bytes for spool and other system defined temporary table space",0,teradata,allocspool,
+teradata.disk_space.alloc_temp.total,gauge,,byte,,"current total allocation in bytes for temporary tables space",0,teradata,alloctemp,
+teradata.disk_space.perm_skew.total,gauge,,,percent,"permissible skew limit percent for permanent space usage at the AMP level",0,teradata,perkskew,
+teradata.disk_space.spool_skew.total,gauge,,,percent,"permissible skew limit percent for spool space usage at the AMP level",0,teradata,spoolskew,
+teradata.disk_space.temp_skew.total,gauge,,,percent,"permissible skew limit percent for temporary space usage at the AMP level",0,teradata,tempskew,
 teradata.disk_space.max_perm,gauge,,byte,,"Maximum permanent space, in bytes, that is allocated to the database on a specified AMP",0,teradata,maxperm,
 teradata.disk_space.max_spool,gauge,,byte,,"Maximum spool space, in bytes, that is allocated to the database on a specified AMP",0,teradata,maxspool,
 teradata.disk_space.max_temp,gauge,,byte,,"Maximum temporary space, in bytes, that is allocated to the database on a specified AMP",0,teradata,maxtemp,
