+++ teradata/tests/fixtures/allspacev.csv
@@ -1,10 +1,10 @@
-0,AdventureWorksDW,$M00ADMR,DimCustomer,0,0,0,1314816,0,0,0,1314816,0,0,0,0,0,0,0,0,0,0,0
-0,AdventureWorksDW,$M00ADMR,DimGeography,0,0,0,36864,0,0,0,36864,0,0,0,0,0,0,0,0,0,0,0
-0,AdventureWorksDW,$M00ADMR,DimOrganization,0,0,0,12288,0,0,0,12288,0,0,0,0,0,0,0,0,0,0,0
-0,AdventureWorksDW,$M00ADMR,AdventureWorksDWBuildVersion,0,0,0,8192,0,0,0,8192,0,0,0,0,0,0,0,0,0,0,0
-0,AdventureWorksDW,$M00ADMR,DimDate,0,0,0,126976,0,0,0,126976,0,0,0,0,0,0,0,0,0,0,0
-0,AdventureWorksDW,$M00ADMR,DimEmployee,0,0,0,4685824,0,0,0,4685824,0,0,0,0,0,0,0,0,0,0,0
-0,AdventureWorksDW,$M00ADMR,DimSalesReason,0,0,0,12288,0,0,0,12288,0,0,0,0,0,0,0,0,0,0,0
-0,AdventureWorksDW,$M00ADMR,DimScenario,0,0,0,8192,0,0,0,8192,0,0,0,0,0,0,0,0,0,0,0
-0,AdventureWorksDW,$M00ADMR,DimAccount,0,0,0,16384,0,0,0,16384,0,0,0,0,0,0,0,0,0,0,0
-0,AdventureWorksDW,$M00ADMR,DimDepartmentGroup,0,0,0,8192,0,0,0,8192,0,0,0,0,0,0,0,0,0,0,0
+0,$M00ADMR,AdventureWorksDW,DimDate,0,0,0,126976,0,0,0,126976,0,0,0,0,0,0,0,0,0,0,0
+0,$M00ADMR,AdventureWorksDW,DimSalesReason,0,0,0,12288,0,0,0,12288,0,0,0,0,0,0,0,0,0,0,0
+0,$M00ADMR,AdventureWorksDW,DimScenario,0,0,0,8192,0,0,0,8192,0,0,0,0,0,0,0,0,0,0,0
+0,$M00ADMR,AdventureWorksDW,All,250000000,19831497687,19831497687,34127872,0,0,0,34127872,0,0,0,,,250000000,19831497687,19831497687,0,0,0
+0,$M00ADMR,AdventureWorksDW,DimCustomer,0,0,0,1314816,0,0,0,1314816,0,0,0,0,0,0,0,0,0,0,0
+0,$M00ADMR,AdventureWorksDW,DimDepartmentGroup,0,0,0,8192,0,0,0,8192,0,0,0,0,0,0,0,0,0,0,0
+0,$M00ADMR,AdventureWorksDW,DimGeography,0,0,0,36864,0,0,0,36864,0,0,0,0,0,0,0,0,0,0,0
+0,$M00ADMR,AdventureWorksDW,DimOrganization,0,0,0,12288,0,0,0,12288,0,0,0,0,0,0,0,0,0,0,0
+0,$M00ADMR,AdventureWorksDW,DimCurrency,0,0,0,12288,0,0,0,12288,0,0,0,0,0,0,0,0,0,0,0
+0,$M00ADMR,AdventureWorksDW,DimEmployee,0,0,0,4685824,0,0,0,4685824,0,0,0,0,0,0,0,0,0,0,0
