+++ teradata/tests/fixtures/ampusagev.csv
@@ -1,14 +1,14 @@
 1,DBC,DBADMIN,7.175999999999999,22943.0,288.04463562011716
-0,DBC,DATADOG,1438.779999999861,9288279.0,57752.6283218328
+1,DBC,DATADOG,1473.4039999998713,9702489.0,59142.4356607004
 0,DBC,DBADMIN,6.98,22182.0,280.17719573974614
-1,DBC,DATADOG,1408.5479999998834,8851378.0,56539.11586028585
-1,DBC,TDWM,314.93199999999894,0.0,12641.370287780719
-0,,DBC,0.036000000000000004,35.0,1.445039978027344
-0,DBC,TDWM,303.91200000000066,0.0,12199.027494506863
+0,DBC,DATADOG,1502.1559999998476,10166877.0,60296.54092315062
+1,DBC,TDWM,315.88400000000024,0.0,12679.583567199717
 1,,DBC,0.07600000000000001,55.0,3.0506399536132816
+0,DBC,TDWM,304.82400000000223,0.0,12235.635173950284
+0,,DBC,0.036000000000000004,35.0,1.445039978027344
 1,DBC,SYSDBA,0.34,1403.0,13.64759979248047
-1,DBC,DBC,2193.987999999999,1709794.0,88066.67698089595
-0,DBC,SYSDBA,0.496,1944.0,19.909439697265626
 0,DBC,DBC,2190.4759999999983,1678380.0,87925.70530303949
+0,DBC,SYSDBA,0.496,1944.0,19.909439697265626
+1,DBC,DBC,2193.987999999999,1709794.0,88066.67698089595
 1,$M00ADMR,BOB,210.08800000000002,178909.0,8432.932191772461
 0,$M00ADMR,BOB,210.56399999999996,193334.0,8452.038831481932
