+++ teradata/tests/fixtures/diskspacev.csv
@@ -1,2 +1,2 @@
-1,AdventureWorksDW,$M00ADMR,250000000,19831497687,19831497687,34127872,0,0,0,34127872,0,0,0,,,250000000,19831497687,19831497687,0,0,0
-0,AdventureWorksDW,$M00ADMR,250000000,19831497687,19831497687,34127872,0,0,0,34127872,0,0,0,,,250000000,19831497687,19831497687,0,0,0
+0,$M00ADMR,AdventureWorksDW,250000000,19831497687,19831497687,34127872,0,0,0,34127872,0,0,0,,,250000000,19831497687,19831497687,0,0,0
+1,$M00ADMR,AdventureWorksDW,250000000,19831497687,19831497687,34127872,0,0,0,34127872,0,0,0,,,250000000,19831497687,19831497687,0,0,0
